Issue #640: move gen_rand funcs to preprocess
