+++ deepreg/model/layer_util.py
@@ -347,161 +347,6 @@ def resample(
     return sampled
 
 
-def gen_rand_affine_transform(
-    batch_size: int, scale: float, seed: (int, None) = None
-) -> tf.Tensor:
-    """
-    Function that generates a random 3D transformation parameters for a batch of data.
-
-    for 3D coordinates, affine transformation is
-
-    .. code-block:: text
-
-        [[x' y' z' 1]] = [[x y z 1]] * [[* * * 0]
-                                        [* * * 0]
-                                        [* * * 0]
-                                        [* * * 1]]
-
-    where each * represents a degree of freedom,
-    so there are in total 12 degrees of freedom
-    the equation can be denoted as
-
-        new = old * T
-
-    where
-
-    - new is the transformed coordinates, of shape (1, 4)
-    - old is the original coordinates, of shape (1, 4)
-    - T is the transformation matrix, of shape (4, 4)
-
-    the equation can be simplified to
-
-    .. code-block:: text
-
-        [[x' y' z']] = [[x y z 1]] * [[* * *]
-                                      [* * *]
-                                      [* * *]
-                                      [* * *]]
-
-    so that
-
-        new = old * T
-
-    where
-
-    - new is the transformed coordinates, of shape (1, 3)
-    - old is the original coordinates, of shape (1, 4)
-    - T is the transformation matrix, of shape (4, 3)
-
-    Given original and transformed coordinates,
-    we can calculate the transformation matrix using
-
-        x = np.linalg.lstsq(a, b)
-
-    such that
-
-        a x = b
-
-    In our case,
-
-    - a = old
-    - b = new
-    - x = T
-
-    To generate random transformation,
-    we choose to add random perturbation to corner coordinates as follows:
-    for corner of coordinates (x, y, z), the noise is
-
-        -(x, y, z) .* (r1, r2, r3)
-
-    where ri is a random number between (0, scale).
-    So
-
-        (x', y', z') = (x, y, z) .* (1-r1, 1-r2, 1-r3)
-
-    Thus, we can directly sample between 1-scale and 1 instead
-
-    We choose to calculate the transformation based on
-    four corners in a cube centered at (0, 0, 0).
-    A cube is shown as below, where
-
-    - C = (-1, -1, -1)
-    - G = (-1, -1, 1)
-    - D = (-1, 1, -1)
-    - A = (1, -1, -1)
-
-    .. code-block:: text
-
-                    G — — — — — — — — H
-                  / |               / |
-                /   |             /   |
-              /     |           /     |
-            /       |         /       |
-          /         |       /         |
-        E — — — — — — — — F           |
-        |           |     |           |
-        |           |     |           |
-        |           C — — | — — — — — D
-        |         /       |         /
-        |       /         |       /
-        |     /           |     /
-        |   /             |   /
-        | /               | /
-        A — — — — — — — — B
-
-    :param batch_size: int
-    :param scale: a float number between 0 and 1
-    :param seed: control the randomness
-    :return: shape = (batch, 4, 3)
-    """
-
-    assert 0 <= scale <= 1
-    np.random.seed(seed)
-    noise = np.random.uniform(1 - scale, 1, [batch_size, 4, 3])  # shape = (batch, 4, 3)
-
-    # old represents four corners of a cube
-    # corresponding to the corner C G D A as shown above
-    old = np.tile(
-        [[[-1, -1, -1, 1], [-1, -1, 1, 1], [-1, 1, -1, 1], [1, -1, -1, 1]]],
-        [batch_size, 1, 1],
-    )  # shape = (batch, 4, 4)
-    new = old[:, :, :3] * noise  # shape = (batch, 4, 3)
-
-    theta = np.array(
-        [np.linalg.lstsq(old[k], new[k], rcond=-1)[0] for k in range(batch_size)]
-    )  # shape = (batch, 4, 3)
-
-    return tf.cast(theta, dtype=tf.float32)
-
-
-def gen_rand_ddf(
-    batch_size: int,
-    image_size: tuple,
-    field_strength: (tuple, list),
-    low_res_size: (tuple, list),
-    seed: (int, None) = None,
-) -> tf.Tensor:
-    """
-    Function that generates a random 3D DDF for a batch of data.
-
-    :param batch_size:
-    :param image_size:
-    :param field_strength: maximum field strength, computed as a U[0,field_strength]
-    :param low_res_size: low_resolution deformation field that will be upsampled to
-        the original size in order to get smooth and more realistic fields.
-    :param seed: control the randomness
-    :return:
-    """
-
-    np.random.seed(seed)
-    low_res_strength = np.random.uniform(0, field_strength, (batch_size, 1, 1, 1, 3))
-    low_res_field = low_res_strength * np.random.randn(
-        batch_size, low_res_size[0], low_res_size[1], low_res_size[2], 3
-    )
-    high_res_field = resize3d(low_res_field, image_size)
-    return high_res_field
-
-
 def warp_grid(grid: tf.Tensor, theta: tf.Tensor) -> tf.Tensor:
     """
     Perform transformation on the grid.
